Alias: "^o" -> fzf-edit-file.
Creating an alias like this in zsh is weird. The way I saw in done in
various peoples configs was to make your function a zle widget with
`zle -N <function name>` and then mapping with `bindkey <key> <func>`.
But that causes some weird things to happen, and I never actually got
the function to run.

The workaround is to use bindkey with the `-s` flag, which literally
sends the character sequence and executes it. Just like `feedkeys` in
vim. Notice the `` at the end of the command as a literal enter. This
executes the command, finally.